requirements.txt
streamlit
pandas
matplotlib
python-docx
